# openapi.v1.yaml
openapi: 3.1.1
info:
  title: TNPSC Book RAG API
  version: 1.0.0
  summary: Frontend contract for the English-first TNPSC textbook RAG MVP.
  description: |
    This checked-in contract contains implemented and planned v1 operations so the frontend can
    generate types and mocks before every backend route exists. `x-implementation-status` records
    current delivery state. FastAPI's generated OpenAPI must conform to the implemented subset.
servers:
  - url: "{origin}"
    variables:
      origin:
        default: http://127.0.0.1:8000
        description: Deployment-specific API origin.
...
paths:
...
  /v1/capabilities:
    get:
      tags: [capabilities]
      operationId: getCapabilities
      summary: Return client-visible features and runtime limits.
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Runtime capabilities.
          headers:
            Cache-Control:
              schema:
                type: string
              example: no-store
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capabilities"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/catalog/filters:
    get:
      tags: [catalog]
      operationId: getCatalogFilters
      summary: Return search filters derived from active ready documents.
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Available search filters.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogFilters"
  /v1/library:
    get:
      tags: [catalog]
      operationId: getLibrary
      summary: Return flat list of all textbook PDF source documents with metadata.
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: List of library items.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryResponse"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/books:
    get:
      tags: [catalog]
      operationId: listBooks
      summary: List conceptual textbooks.
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - name: standard
          in: query
          description: Repeatable textbook-standard filter.
          schema:
            type: array
            maxItems: 5
            uniqueItems: true
            items:
              $ref: "#/components/schemas/TextbookStandard"
          style: form
          explode: true
        - name: subject
          in: query
          description: Repeatable canonical subject filter.
          schema:
            type: array
            maxItems: 20
            uniqueItems: true
            items:
              type: string
              minLength: 1
              maxLength: 200
          style: form
          explode: true
        - name: q
          in: query
          schema:
            type: string
            minLength: 1
            maxLength: 200
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/IncludeCount"
      responses:
        "200":
          description: Paginated books.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookPage"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    post:
      tags: [catalog]
      operationId: createBook
      summary: Create a conceptual textbook.
      x-implementation-status: implemented
      security:
        - HTTPBearer: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookRequest"
      responses:
        "201":
          description: Book created.
          headers:
            Location:
              $ref: "#/components/headers/Location"
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/books/{book_id}:
    parameters:
      - $ref: "#/components/parameters/BookId"
    get:
      tags: [catalog]
      operationId: getBook
      summary: Get a book and its registered documents.
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Book detail.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/books/{book_id}/documents:
    parameters:
      - $ref: "#/components/parameters/BookId"
    post:
      tags: [catalog, ingestion]
      operationId: uploadBookDocument
      summary: Upload and queue a digital textbook PDF.
      x-implementation-status: implemented
      security:
        - HTTPBearer: []
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DocumentUploadRequest"
      responses:
        "202":
          description: Upload stored and ingestion queued.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentUploadAccepted"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token issued by the configured OpenID Connect provider.
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      schema:
        $ref: "#/components/schemas/RequestIdentifier"
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: "^[A-Za-z0-9._:-]+$"
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        minLength: 1
    IncludeCount:
      name: include_count
      in: query
      required: false
      schema:
        type: boolean
        default: false
    BookId:
      name: book_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
    RequestId:
      required: true
      schema:
        $ref: "#/components/schemas/RequestIdentifier"
    Location:
      required: true
      schema:
        type: string
  responses:
    NotFound:
      description: Resource was not found.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: Request conflicts with current state or idempotency history.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    PayloadTooLarge:
      description: Upload exceeds the configured byte limit.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    UnsupportedMediaType:
      description: Upload is not a supported PDF.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unprocessable:
      description: Validation, cursor, or document-content error.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    ServiceUnavailable:
      description: A required service is temporarily unavailable.
      headers:
        Retry-After:
          schema:
            oneOf:
              - type: integer
                minimum: 0
              - type: string
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    RequestIdentifier:
      type: string
      minLength: 1
      maxLength: 200
      description: Accepted or generated safe printable correlation identifier.
    TextbookStandard:
      type: integer
      enum: [6, 7, 8, 9, 10]
    DocumentLanguage:
      type: string
      enum: [english]
    DocumentState:
      type: string
      enum: [uploaded, queued, extracting, chunking, embedding, ready, failed]
    CatalogStatus:
      type: string
      enum: [empty, processing, ready, failed]
      description: >-
        Derived catalog availability. A book remains ready while a replacement document is
        processing or has failed if an older active document is still ready.
...
    CapabilityFeatures:
      type: object
      required:
        - catalog_mutation
        - ingestion_inspection
        - semantic_search
        - answer_generation
        - answer_streaming
        - answer_recovery
      properties:
        catalog_mutation: {type: boolean}
        ingestion_inspection: {type: boolean}
        semantic_search: {type: boolean}
        answer_generation: {type: boolean}
        answer_streaming: {type: boolean}
        answer_recovery: {type: boolean}
      additionalProperties: false
    CapabilityLimits:
      type: object
      required:
        - max_upload_bytes
        - max_query_characters
        - max_top_k
        - max_answer_characters_per_section
        - answer_timeout_seconds
        - answer_retention_seconds
        - thumbnail_max_edge_pixels
      properties:
        max_upload_bytes: {type: integer, minimum: 1}
        max_query_characters: {type: integer, minimum: 1}
        max_top_k: {type: integer, minimum: 1}
        max_answer_characters_per_section: {type: integer, minimum: 1}
        answer_timeout_seconds: {type: integer, minimum: 1}
        answer_retention_seconds: {type: integer, minimum: 1}
        thumbnail_max_edge_pixels: {type: integer, minimum: 1}
      additionalProperties: false
    UploadCapabilities:
      type: object
      required: [accepted_media_types, requires_text_layer]
      properties:
        accepted_media_types:
          type: array
          minItems: 1
          items: {type: string}
        requires_text_layer: {type: boolean}
      additionalProperties: false
    Capabilities:
      type: object
      required: [api_version, features, limits, upload]
      properties:
        api_version:
          type: string
          const: v1
        features:
          $ref: "#/components/schemas/CapabilityFeatures"
        limits:
          $ref: "#/components/schemas/CapabilityLimits"
        upload:
          $ref: "#/components/schemas/UploadCapabilities"
      additionalProperties: false
    ValidationFieldError:
      type: object
      required: [field, message, code]
      properties:
        field:
          type: string
          description: >-
            Stable path such as body.filters.standards.0, query.limit, path.book_id,
            header.Idempotency-Key, or form.file.
        message: {type: string}
        code: {type: string}
      additionalProperties: false
    Problem:
      type: object
      required: [type, title, status, detail, instance, code, request_id, errors]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string}
        instance: {type: string, format: uri-reference}
        code: {type: string}
        request_id:
          $ref: "#/components/schemas/RequestIdentifier"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ValidationFieldError"
      additionalProperties: false
...
    Book:
      type: object
      required:
        - id
        - title
        - standard
        - subject
        - language
        - publisher
        - catalog_identifier
        - catalog_status
        - document_count
        - active_document_id
        - latest_document_id
        - latest_document_state
        - created_at
        - updated_at
      properties:
        id: {type: string, format: uuid}
        title: {type: string, minLength: 1, maxLength: 500}
        standard:
          $ref: "#/components/schemas/TextbookStandard"
        subject: {type: string, minLength: 1, maxLength: 200}
        language:
          $ref: "#/components/schemas/DocumentLanguage"
        publisher: {type: string, minLength: 1, maxLength: 300}
        catalog_identifier: {type: [string, "null"], minLength: 1, maxLength: 200}
        catalog_status:
          $ref: "#/components/schemas/CatalogStatus"
        document_count: {type: integer, minimum: 0}
        active_document_id:
          type: [string, "null"]
          format: uuid
        latest_document_id:
          type: [string, "null"]
          format: uuid
        latest_document_state:
          oneOf:
            - $ref: "#/components/schemas/DocumentState"
            - type: "null"
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    CreateBookRequest:
      type: object
      required: [title, standard, subject, publisher]
      properties:
        title: {type: string, minLength: 1, maxLength: 500}
        standard:
          $ref: "#/components/schemas/TextbookStandard"
        subject: {type: string, minLength: 1, maxLength: 200}
        language:
          allOf:
            - $ref: "#/components/schemas/DocumentLanguage"
          default: english
        publisher: {type: string, minLength: 1, maxLength: 300}
        catalog_identifier: {type: [string, "null"], minLength: 1, maxLength: 200}
      additionalProperties: false
...
    BookDetail:
      allOf:
        - $ref: "#/components/schemas/Book"
        - type: object
          required: [documents]
          properties:
            documents:
              type: array
              items:
                $ref: "#/components/schemas/DocumentSummary"
    LibraryItem:
      type: object
      required:
        - document_id
        - book_id
        - title
        - standard
        - subject
        - edition
        - publisher
        - source_filename
        - file_size_bytes
        - state
        - page_count
        - uploaded_at
        - active
      properties:
        document_id: {type: string, format: uuid}
        book_id: {type: string, format: uuid}
        title: {type: string}
        standard:
          $ref: "#/components/schemas/TextbookStandard"
        subject: {type: string}
        edition: {type: string}
        publisher: {type: string}
        source_filename: {type: string}
        file_size_bytes: {type: integer, minimum: 1}
        state:
          $ref: "#/components/schemas/DocumentState"
        page_count: {type: [integer, "null"], minimum: 1}
        uploaded_at: {type: string, format: date-time}
        active: {type: boolean}
    LibraryResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/LibraryItem"
...
    PaginationMetadata:
      type: object
      required: [previous_cursor, next_cursor, total_items]
      properties:
        previous_cursor: {type: [string, "null"]}
        next_cursor: {type: [string, "null"]}
        total_items: {type: [integer, "null"], minimum: 0}
      description: >-
        total_items is non-null only when include_count=true. Cursors support adjacent navigation;
        arbitrary page-number jumps are not supported.
    BookPage:
      allOf:
        - $ref: "#/components/schemas/PaginationMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Book"
...
    DocumentUploadRequest:
      type: object
      required: [file, edition]
      properties:
        file:
          type: string
          format: binary
        edition:
          type: string
          minLength: 1
          maxLength: 200
      additionalProperties: false
    UploadLinks:
      type: object
      required: [document, ingestion_run]
      properties:
        document: {type: string, pattern: "^/v1/documents/"}
        ingestion_run: {type: string, pattern: "^/v1/ingestion-runs/"}
      additionalProperties: false
    DocumentUploadAccepted:
      type: object
      required: [document, ingestion_run, poll_after_seconds, links]
      properties:
        document:
          $ref: "#/components/schemas/DocumentSummary"
        ingestion_run:
          $ref: "#/components/schemas/IngestionRun"
        poll_after_seconds: {type: integer, minimum: 1}
        links:
          $ref: "#/components/schemas/UploadLinks"
      additionalProperties: false
...
    CatalogBookOption:
      type: object
      required: [id, title, standard, subject]
      properties:
        id: {type: string, format: uuid}
        title: {type: string}
        standard:
          $ref: "#/components/schemas/TextbookStandard"
        subject: {type: string}
    CatalogFilters:
      type: object
      required: [standards, subjects, books]
      properties:
        standards:
          type: array
          items:
            $ref: "#/components/schemas/TextbookStandard"
        subjects:
          type: array
          items: {type: string}
        books:
          type: array
          items:
            $ref: "#/components/schemas/CatalogBookOption"
      additionalProperties: false
...